+++ src/deepuq/methods/__init__.py
@@ -1,14 +1,16 @@
+from .ensembles import DeepEnsembleWrapper
+from .laplace import LaplaceWrapper
 from .mc_dropout import MCDropoutWrapper
-from .vi import BayesByBackpropMLP, predict_vi_uq, vi_elbo_step
 from .mcmc import (
     SGLDOptimizer,
     collect_posterior_samples,
     predict_with_samples,
     predict_with_samples_uq,
 )
-from .laplace import LaplaceWrapper
+from .vi import BayesByBackpropMLP, predict_vi_uq, vi_elbo_step
 
 __all__ = [
+    "DeepEnsembleWrapper",
     "MCDropoutWrapper",
     "BayesByBackpropMLP",
     "predict_vi_uq",